move `liveness_map` into the `liveness` module
